+++ R/parse-iupac-condensed.R
@@ -36,9 +36,8 @@
 #'
 #' @export
 parse_iupac_condensed <- function(x, on_failure = "error", progress = FALSE) {
-  struc_parser_wrapper(
+  normalized_struc_parser_wrapper(
     x,
-    do_parse_iupac_condensed,
     on_failure = on_failure,
     progress = progress
   )